requirements.txt
# Python dependencies
Pillow
